Refactoring: - Move BMP-specific functionality into new subclass of Bitmap - Adapt using code and example
