+++ examples/Drawing/index.js
@@ -5,7 +5,7 @@ const bitmapManipulation = require("../../");
 let targetColorDepth = 4;
 
 // Create bitmap with 1 byte per pixel to draw on
-let bitmap = new bitmapManipulation.Bitmap(400, 300);
+let bitmap = new bitmapManipulation.BMPBitmap(400, 300);
 let palette = bitmap.getPalette();
 
 // Fill image with light grey background
@@ -22,20 +22,20 @@ bitmap.drawEllipse(130, 70, 50, 40, null, palette.indexOf(0x008000));
 bitmap.drawEllipse(190, 70, 50, 50, palette.indexOf(0xff0000), palette.indexOf(0x000080), 3);
 
 // Draw overlay bitmap with transparent pixels
-let overlayBitmap = bitmapManipulation.Bitmap.fromFile("Overlay.bmp");
-bitmap.drawBitmap(overlayBitmap, 10, 130, overlayBitmap.getPalette().indexOf(0xff00ff/*magenta*/));
+let overlayBitmap = bitmapManipulation.BMPBitmap.fromFile("Overlay.bmp");
+bitmap.drawBitmap(overlayBitmap, 10, 130, overlayBitmap.getPalette().indexOf(0xff00ff));
 
 // Draw text
 let font = new bitmapManipulation.Font("../Font/Arial.json");
 font.setSize(20);
-font.setColor(font.getBitmap().getPalette().indexOf(0x000000/*black*/));
+font.setColor(font.getBitmap().getPalette().indexOf(0x000000));
 bitmap.drawText(font, "The quick brown fox\njumps over the lazy dog.\n3.14159265358979323846",
 		10, 190);
 
 bitmap.changeColorDepth(targetColorDepth);
 
 // Draw gradient rectangle
-overlayBitmap = new bitmapManipulation.Bitmap(50, 50);
+overlayBitmap = new bitmapManipulation.BMPBitmap(50, 50);
 let overlayBitmapPalette = overlayBitmap.getPalette();
 overlayBitmapPalette.length = 0;
 for (let i = 0; i <= 0xff; i++) {
