+++ index.js
@@ -5,4 +5,5 @@
  */
 exports.Endianness = require("./lib/Endianness");
 exports.Bitmap = require("./lib/Bitmap");
+exports.BMPBitmap = require("./lib/BMPBitmap");
 exports.Font = require("./lib/Font");
