+++ lib/Font.js
@@ -3,7 +3,7 @@
 const assert = require("assert");
 const fs = require("fs");
 const path = require("path");
-const Bitmap = require("./Bitmap");
+const BMPBitmap = require("./BMPBitmap");
 
 /**
  * Creates a bitmap font that's read from JSON and image files. There's a main JSON file that
@@ -71,7 +71,7 @@ module.exports = class Font {
             detailsAndBitmap.details.kernings = newKernings;
             // Create bitmap from font image file
             assert(detailsAndBitmap.details.pages.length >= 1);
-            detailsAndBitmap.bitmap = Bitmap.fromFile(
+            detailsAndBitmap.bitmap = BMPBitmap.fromFile(
                 path.join(path.dirname(filePath), detailsAndBitmap.details.pages[0]));
             let palette = detailsAndBitmap.bitmap.getPalette();
             let currentColor = palette.indexOf(0x000000/*black*/);
